requirements.txt
# 依赖包列表
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-crontab>=3.0.0
jinja2>=3.1.0
python-dateutil>=2.8.0
